+++ src/tests/ion_hash/mod.rs
@@ -1,0 +1,15 @@
+pub mod annotations;
+pub mod blob;
+pub mod bool;
+pub mod clob;
+pub mod decimal;
+pub mod float;
+pub mod integer;
+pub mod list;
+pub mod null;
+pub mod sexp;
+pub mod string;
+pub mod r#struct;
+pub mod symbol;
+pub mod timestamp;
+pub mod general;
